Level2MostlyDone
tut-> level2 "works". Gates are an issue, I think its the gate timer stuff. This also hinders level3 from loading. Also need to implement third mineral. and the second for level 2. and connect all the gates and levers and stuff. Transitions are fine for the most part up until level 3. Haven't tested it out yet since I would need to mess with a lot of code I don't know the improtance of. If you;re reading this yo.  What's the difference between an enzyme and a hormone??? You can't hear an enzyme