Back to Beg. Progress (read desc for details)
Once the game ends, it now goes back to the main menu. However, when you press play again, it's just a black screen. I suspect theres some canvas/color rect silly business happening here. Also as a heads up, For the game not to crash, I only remove the player node if the game over variable is true. Meaning once the main menu pops back up, the player is still existing somewhere. It'll be easier to fix that once we figure out the black screen problems